Fixing some problems detected with clang-tidy

+++ src/driving_distance/many_to_dist_driving_distance.c
@@ -48,7 +48,7 @@ static
 void process(
         char* edges_sql,
         ArrayType *starts,
-        float8 distance,
+        double distance,
         bool directed,
         bool equicost,
         MST_rt **result_tuples,
